refactor: axios组件结构调整，代码优化
