+++ packages/axios/src/config.default.ts
@@ -1,0 +1,6 @@
+/**
+ * 默认配置
+ */
+export default {
+  axios: {},
+};
