+++ packages/axios/src/index.ts
@@ -2,12 +2,12 @@ import axios from 'axios';
 import * as Axios from 'axios';
 
 export { AxiosConfiguration as Configuration } from './configuration';
-export * from './interface';
-export * from './serviceManager';
+export * from './http-service.factory';
+export * from './http-service';
 export {
   /**
    * @deprecated Use `Axios` directly
    */
   axios,
+  Axios,
 };
-export { Axios };
